backend/requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==42.0.8
dnspython==2.8.0
eventlet==0.41.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.6.1
Flask-SQLAlchemy==3.1.1
greenlet==3.4.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycparser==3.0
PyJWT==2.12.1
PyMySQL==1.1.1
python-dotenv==1.0.1
python-engineio==4.13.1
python-socketio==5.16.1
requests==2.31.0
simple-websocket==1.1.0
SQLAlchemy==2.0.49
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.0.3
wsproto==1.3.2
